refactor(web): optimize Dify Builder streaming updates
