+++ web/app/components/workflow-app/components/dify-builder/conversation/group-conversation-items.ts
@@ -1,4 +1,4 @@
-import type { ConversationItem } from '@/app/components/dify-builder/types'
+import type { ConversationItem } from '../types'
 
 export type DifyBuilderConversationGroup =
   | { type: 'standalone'; item: ConversationItem; invalidated: false }
