+++ web/app/components/workflow-app/components/dify-builder/session/projection.ts
@@ -1,0 +1,39 @@
+import type { DifyBuilderCommitEventData } from '@dify/contracts/api/console/dify-builder/types.gen'
+import type { ConversationItem, SessionView } from '../types'
+
+export const isTerminalView = (view: SessionView) =>
+  view.run_status === 'complete' || view.run_status === 'failed'
+
+const mergeConversation = (
+  current: ConversationItem[],
+  committed: ConversationItem[],
+): ConversationItem[] => {
+  const bySequence = new Map(current.map((item) => [item.seq, item]))
+  committed.forEach((item) => bySequence.set(item.seq, item))
+  return [...bySequence.values()].sort((left, right) => left.seq - right.seq)
+}
+
+export const projectSessionView = (
+  current: SessionView | null,
+  next: SessionView,
+): SessionView | null => {
+  if (current?.session_id === next.session_id && next.version < current.version) return null
+  return next
+}
+
+export const projectCommit = (
+  current: SessionView | null,
+  commit: DifyBuilderCommitEventData,
+): SessionView | null => {
+  if (!current || current.session_id !== commit.session_id || commit.version <= current.version)
+    return null
+
+  return {
+    ...current,
+    version: commit.version,
+    state: commit.state,
+    canvas_read_only: true,
+    run_status: 'executing',
+    conversation: mergeConversation(current.conversation, commit.items),
+  }
+}
