+++ web/app/components/workflow/header/checklist/index.tsx
@@ -1,6 +1,6 @@
 import type { ChecklistItem } from '../../hooks/use-checklist'
 import type { CommonEdgeType } from '../../types'
-import type { ChecklistErrorPayload } from '@/app/components/dify-builder/types'
+import type { ChecklistErrorPayload } from '@/app/components/workflow-app/components/dify-builder/types'
 import { cn } from '@langgenius/dify-ui/cn'
 import {
   Popover,
